requirements.txt.txt
PySide6
requests
openai
openai-whisper
pysubs2